feat: refactor build-plugin-fusion
