--- packages/build-scripts-config/CHANGELOG.md
@@ -1,45 +1,0 @@
-# Changelog
-
-## 3.0.3
-
-- [feat] support webp image file
-
-## 3.0.2
-
-- [fix] lock sass version until it's define a js api for handling warnings
-
-## 3.0.1
-
-- [feat] enable stage 3 for postcss preset env
-
-## 3.0.0
-
-- [feat] upgrate less and less-loader
-
-## 2.0.0
-
-- [feat] remove dependencies of node-sass
-
-## 1.0.0
-
-- [feat] refactor config with @builder/babel-preset-ice
-
-## 0.1.8
-
-- [fix] support css modules hmr
-
-## 0.1.7
-
-- [fix] use `config.path` to set postcss options [#743026](http://gitlab.alibaba-inc.com/a-lib/build-scripts/merge_requests/743026)
-
-## 0.1.6
-
-- [fix] files created before watching start make compile twice
-
-## 0.1.5
-
-- [fix] default option of Terser plugin
-
-## 0.1.4
-
-- [feat] use Terser plugin instead of Uglify plugin.
