+++ packages/plugin-fusion-legacy/package.json
@@ -10,6 +10,22 @@
     "fusion",
     "build-scripts"
   ],
+  "dependencies": {
+    "@builder/app-helpers": "^2.1.0",
+    "babel-plugin-import": "^1.11.2",
+    "ice-skin-loader": "^0.3.0",
+    "postcss": "^7.0.32",
+    "postcss-plugin-rpx2vw": "^0.0.2",
+    "resolve-sass-import": "^0.1.0",
+    "webpack-sources": "^2.0.0"
+  },
+  "devDependencies": {
+    "@types/webpack-sources": "^2.0.0"
+  },
+  "peerDependencies": {
+    "sass": "^1.0.0",
+    "webpack": "^4.0.0"
+  },
   "author": "",
   "license": "MIT"
 }
