+++ packages/plugin-fusion/package.json
@@ -14,21 +14,10 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@builder/app-helpers": "^2.1.0",
-    "babel-plugin-import": "^1.11.2",
-    "babel-plugin-module-resolver": "^4.0.0",
-    "find-root": "^1.1.0",
-    "ice-skin-loader": "^0.3.0",
     "lodash": "^4.17.15",
-    "postcss": "^7.0.32",
-    "postcss-plugin-rpx2vw": "^0.0.2",
-    "resolve-sass-import": "^0.1.0",
-    "semver": "^6.1.0",
-    "webpack-plugin-import": "^0.2.5",
-    "webpack-sources": "^2.0.0"
+    "webpack-plugin-import": "^0.2.5"
   },
   "devDependencies": {
-    "@types/webpack-sources": "^2.0.0",
     "@types/lodash": "^4.0.0"
   },
   "peerDependencies": {
