+++ packages/plugin-fusion/src/index.ts
@@ -1,359 +1,49 @@
-import * as fs from 'fs';
-import * as path from 'path';
 import { upperFirst, camelCase } from 'lodash';
 import { IPlugin } from 'build-scripts';
-import AppendStyleWebpackPlugin from './webpackPlugins/appendStyleWebpackPlugin';
 
 interface PluginOptions {
   externalNext: boolean;
   cssVariable: boolean;
-  themePackage: string;
-  nextLibDir: string;
   importOptions: Partial<{libraryDirectory: string; style: any}>
 }
 
-function normalizeEntry(entry, preparedChunks) {
-  const preparedName = (preparedChunks || [])
-    .filter((module) => {
-      return typeof module.name !== 'undefined';
-    })
-    .map((module) => module.name);
-
-  return Object.keys(entry).concat(preparedName);
+interface BabelOptions {
+  plugins?: [string, ...any][];
+  presets?: [string, any][];
 }
 
-function replaceBlank(str) {
-  return str.replace(/[\r\n ]/g, '').replace(/\\/g, '\\\\');
-}
-
-function addCSSVariableCode({ rootDir, themesCssVars, defaultTheme, cssVariable, config, log }) {
-  try {
-    const tempDir = path.join(rootDir, './node_modules');
-    const jsPath = path.join(tempDir, 'change-theme.js');
-    fs.writeFileSync(jsPath, getThemeCode(themesCssVars, defaultTheme, cssVariable));
-
-    // add theme.js to entry
-    const entryNames = Object.keys(config.entryPoints.entries());
-    entryNames.forEach((name) => {
-      config.entry(name).prepend(jsPath);
-    });
-  } catch (err) {
-    log.error('fail to add theme.js to entry');
-    log.error(err);
-  }
-}
-
-function getVariablesPath({
-  packageName, filename = 'variables.scss', slient = false
-}) {
-  let filePath = '';
-  const variables = `${packageName}/${filename}`;
-  try {
-    filePath = require.resolve(variables);
-  } catch (err) {
-    if (!slient) {
-      console.log('[ERROR]', `fail to resolve ${variables}`);
-    }
-  }
-  return filePath;
-}
-
-const plugin: IPlugin = ({ onGetWebpackConfig, log, context }, options) => {
-
-};
-
-export default plugin;
-
-module.exports = async ({ onGetWebpackConfig, log, context, getAllTask }, plugionOptions = {}) => {
-  const {
-    themePackage,
-    themeConfig,
-    nextLibDir = 'es',
-    usePx2Vw = false,
-    px2vwOptions = {},
-    style = true,
-    uniteNextLib,
-    externalNext,
-    importOptions = {},
-    componentOptions = {},
-    enableColorNames,
-    cssVariable,
-  } = plugionOptions;
-  let { uniteBaseComponent } = plugionOptions;
-  const { rootDir, pkg, userConfig, webpack } = context;
-
+const plugin: IPlugin = ({ onGetWebpackConfig, getAllTask, context }, options) => {
+  const { externalNext, cssVariable = true, importOptions = { style: true, libraryDirectory: 'es' } } = (options || {}) as Partial<PluginOptions>;
+  const { userConfig } = context;
   const taskNames = getAllTask();
   // ignore externals rule and babel-plugin-import when compile dist
   const ignoreTasks = ['component-dist'];
   taskNames.forEach((taskName) => {
-    onGetWebpackConfig(taskName, (config) => {
-      // 1. 支持主题能力
-      if (themePackage) {
-        if (Array.isArray(themePackage)) {
-          log.info('已启用 themePackage 多主题功能');
-        } else {
-          log.info('使用 Fusion 组件主题包：', themePackage);
-        }
-      }
-      if (themeConfig) {
-        log.info('自定义 Fusion 组件主题变量：', themeConfig);
-      }
-  
-      let replaceVars = {};
-      let defaultScssVars = {};
-      let defaultTheme = '';
-      if (Array.isArray(themePackage)) {
-        const themesCssVars = {};
-        const varsPath = !cssVariable && getVariablesPath({ packageName: '@alifd/next', slient: true });
-        // get scss variables and generate css variables
-        themePackage.forEach(({ name, ...themeData }) => {
-          const themePath = getVariablesPath({ packageName: name, filename: `variables.${cssVariable ? 'css' : 'scss'}` });
-          if (!cssVariable) {
-            const configData = themeData.themeConfig || {};
-            let themeVars = {};
-            let calcVars = {};
-            if (varsPath) {
-              calcVars = getCalcVars(varsPath, themePath, configData);
-            }
-            try {
-              themeVars = getThemeVars(themePath, Object.assign({}, calcVars, configData ), enableColorNames);
-            } catch (err) {
-              log.error('get theme variables err:', err);
-            }
-            replaceVars = themeVars.scssVars;
-            defaultScssVars = themeVars.originTheme;
-            themesCssVars[name] = themeVars.cssVars;
-          } else if (themePath) {
-            // read css variables from css file
-            themesCssVars[name] = replaceBlank(fs.readFileSync(themePath, 'utf-8'));
-          }
-
-          if (themeData.default) {
-            defaultTheme = name;
-          }
-        });
-        defaultTheme = defaultTheme || (themePackage[0] && themePackage[0].name);
-        addCSSVariableCode({ defaultTheme, config, themesCssVars, rootDir, cssVariable, log });
-      } else if (cssVariable && themePackage) {
-        // add css variable code when cssVariable is true
-        const cssVariablePath = getVariablesPath({ packageName: themePackage, filename: 'variables.css'});
-        if (cssVariablePath) {
-          const cssVariables = replaceBlank(fs.readFileSync(cssVariablePath, 'utf-8'));
-          addCSSVariableCode({ defaultTheme: themePackage, config, themesCssVars: {[themePackage]: cssVariables}, rootDir, cssVariable, log });
-        }
-      }
-  
-      if (usePx2Vw) {
-        ['css', 'scss', 'scss-module'].forEach(rule => {
-          config.module
-            .rule(rule)
-            .use('postcss-loader')
-            .tap((options) => {
-              const { plugins = [] } = options;
-              return {
-                ...options,
-                plugins: [
-                  ...plugins,
-                  // eslint-disable-next-line
-                  require('postcss-plugin-rpx2vw'),
-                  // eslint-disable-next-line
-                  require('./postcssPlugins/postcssPluginPx2vw')(px2vwOptions),
-                ],
-              };
-            });
-        });
-      };
-
-      const themeFile = typeof themePackage === 'string' && getVariablesPath({ packageName: themePackage, slient: true });
-
-      ['scss', 'scss-module'].forEach((rule) => {
-        config.module
-          .rule(rule)
-          .use('ice-skin-loader')
-          .loader(require.resolve('ice-skin-loader'))
-          .options({
-            themeFile,
-            themeConfig: Object.assign(
-              {},
-              defaultScssVars,
-              replaceVars,
-              themeConfig || {}
-            ),
-          });
-      });
-  
-      // check icons.scss
-      const iconPackage = defaultTheme || themePackage;
-      const iconScssPath = iconPackage && getVariablesPath({ packageName: iconPackage, filename: 'icons.scss', slient: true });
-      if (iconScssPath && fs.existsSync(iconScssPath)) {
-        const appendStylePluginOption = {
-          type: 'sass',
-          srcFile: iconScssPath,
-          variableFile: getVariablesPath({ packageName: iconPackage }),
-          compileThemeIcon: true,
-          themeConfig: themeConfig || {},
-          distMatch: (chunkName, compilerEntry, compilationPreparedChunks) => {
-            const entriesAndPreparedChunkNames = normalizeEntry(
-              compilerEntry,
-              compilationPreparedChunks,
-            );
-            // 仅对 css 的 chunk 做 处理
-            if (entriesAndPreparedChunkNames.length && /\.css$/.test(chunkName)) {
-              // css/index.css -> index css/index.[hash].css -> index
-              // css/_component_.usage.css -> _component_.usage
-              const assetsBaseName = path.basename(chunkName, path.extname(chunkName));
-              const assetsFromEntry = userConfig.hash
-                ? assetsBaseName.substring(0, assetsBaseName.lastIndexOf('.'))
-                : assetsBaseName;
-              if (entriesAndPreparedChunkNames.indexOf(assetsFromEntry) !== -1) {
-                return true;
-              }
-            }
-            return false;
-          },
-        };
-        config.plugin('AppendStyleWebpackPlugin').use(AppendStyleWebpackPlugin, [appendStylePluginOption]);
-      }
-  
-      const crossendBabelLoader = [];
-  
-      if ('componentOptions' in plugionOptions) {
-        const { bizComponent = [], customPath = '', componentMap = {} } = componentOptions;
-        const mixBizCom = {};
-  
-        // bizComponent: ['@alifd/anchor', '@alifd/pro-components'],
-  
-        if (Array.isArray(bizComponent)) {
-          bizComponent.forEach(com => {
-            mixBizCom[com] = `${com}${customPath}`;
-          });
-        }
-  
-        // componentMap: {
-        //  '@alifd/pro-components': '@alifd/pro-components/lib/mobile',
-        //  '@alifd/pro-components-2': '@alifd/pro-components-2-mobile'
-        // }
-        const mapList = Object.keys(componentMap);
-        if (mapList.length > 0) {
-          mapList.forEach(orgName => {
-            mixBizCom[orgName] = componentMap[orgName];
-          });
-        }
-  
-        crossendBabelLoader.push(require.resolve('babel-plugin-module-resolver'), {
-          alias: mixBizCom
-        });
-      }
-      // 2. 组件（包含业务组件）按需加载&样式自动引入
-      // babel-plugin-import: 基础组件
-      // remove babel-plugin-import if external next
+    onGetWebpackConfig((config) => {
       if (!externalNext && !ignoreTasks.includes(taskName)) {
-        const importConfigs = [{
-          libraryName: '@icedesign/base',
-          style,
-        }, {
-          libraryName: '@alife/next',
-          style,
-        }, {
+        const babelPluginImportOptions = {
           libraryName: '@alifd/next',
-          libraryDirectory: nextLibDir,
-          style: cssVariable ? (name) => `${name}/style2` : style,
           ...importOptions,
-        }];
+        };
+        if (cssVariable) {
+          babelPluginImportOptions.style = (name: string) => `${name}/style2`;
+        }
         ['jsx', 'tsx'].forEach((rule) => {
           config.module
             .rule(rule)
             .use('babel-loader')
-            .tap((options) => {
-              const plugins = options.plugins.concat(
-                importConfigs.map((itemConfig) => {
-                  return [require.resolve('babel-plugin-import'), itemConfig, itemConfig.libraryName];
-                }),
+            .tap((babelOptions: BabelOptions) => {
+              const plugins = babelOptions.plugins.concat(
+                [require.resolve('babel-plugin-import'), babelPluginImportOptions, babelPluginImportOptions.libraryName]
               );
-              if (crossendBabelLoader.length > 0) {
-                plugins.push(crossendBabelLoader);
-              }
-              options.plugins = plugins;
-              return options;
+              return {
+                ...babelOptions,
+                plugins,
+              };
             });
         });
       }
-  
-      // 业务组件：不可枚举，使用 webpack-plugin-import，内置逻辑（pkg.componentConfig || pkg.stylePath）
-      // compatible with build-plugin which do not set up WebpackPluginImport
-      if (!config.plugins.get('WebpackPluginImport')) {
-        config.plugin('WebpackPluginImport')
-          .use(WebpackPluginImport, [[
-            // 老的业务组件里没有 stylePath or componentConfig
-            {
-              libraryName: /@ali\/ice-.*/,
-              stylePath: 'style.js',
-            },
-          ]]);
-      }
-  
-      // 3. uniteBaseComponent
-      // uniteBaseComponent: true => @icedesign/base（兼容老的逻辑）
-      // uniteBaseComponent: "@alife/next" => @alife/next
-      uniteBaseComponent = uniteBaseComponent && (
-        typeof uniteBaseComponent === 'string' ? uniteBaseComponent : '@icedesign/base'
-      );
-  
-      if (uniteBaseComponent) {
-        // 统一基础组件包：@ali/ice, @alife/next, @icedesign/base -> @icedesign/base
-        log.info('uniteBaseComponent 开启，基础包统一到：', uniteBaseComponent);
-  
-        const alias = {
-          // @ali/ice -> uniteBaseComponent
-          '@ali/ice/global.scss': `${uniteBaseComponent}/reset.scss`,
-          '@ali/ice/lib/row$': `${uniteBaseComponent}/lib/_components/@alife/next-grid/lib/row.js`,
-          '@ali/ice/lib/col$': `${uniteBaseComponent}/lib/_components/@alife/next-grid/lib/col.js`,
-          // sass 里 @import '~xxx'
-          '@ali/ice/base.scss': `${uniteBaseComponent}/lib/core/index.scss`,
-          '@ali/ice': uniteBaseComponent,
-  
-          // @alife/next -> uniteBaseComponent
-          '@alife/next/lib/_components/@alife/next-core/lib/index.scss': `${uniteBaseComponent}/reset.scss`,
-          '@alife/next/reset.scss': `${uniteBaseComponent}/reset.scss`,
-          // sass 里 @import '~xxx'
-          '@alife/next/variables.scss': `${uniteBaseComponent}/variables.scss`,
-          '@alife/next/lib/core/index.scss': `${uniteBaseComponent}/lib/core/index.scss`,
-          '@alife/next': `${uniteBaseComponent}`,
-  
-          // @icedesign/base -> uniteBaseComponent
-          '@icedesign/base/reset.scss': `${uniteBaseComponent}/reset.scss`,
-          // sass 里 @import '~xxx'
-          '@icedesign/base/variables.scss': `${uniteBaseComponent}/variables.scss`,
-          '@icedesign/base/lib/core/index.scss': `${uniteBaseComponent}/lib/core/index.scss`,
-          '@icedesign/base': `${uniteBaseComponent}`,
-        };
-  
-        config.merge({
-          resolve: {
-            alias,
-          },
-        });
-      }
-  
-      // 4. 检测组件版本
-      config.plugin('CheckIceComponentsDepsPlugin')
-        .use(CheckIceComponentsDepsPlugin, [{
-          pkg,
-          log,
-        }]);
-  
-      if (uniteNextLib) {
-        const replaceRegex = new RegExp(`(alife|alifd)/next/${nextLibDir === 'es' ? 'lib' : 'es'}`);
-        config.plugin('UniteNextLib')
-          .use(webpack.NormalModuleReplacementPlugin, [
-            /@(alife|alifd)\/next\/(.*)/,
-            function(resource) {
-              // eslint-disable-next-line no-param-reassign
-              resource.request = resource.request.replace(replaceRegex, `alifd/next/${nextLibDir}`);
-            },
-          ]);
-      }
-  
+
       if (externalNext && !ignoreTasks.includes(taskName)) {
         const externals = [];
         if (userConfig.externals) {
@@ -362,7 +52,7 @@ module.exports = async ({ onGetWebpackConfig, log, context, getAllTask }, plugio
         const feNextRegex = /@alife\/next\/(es|lib)\/([-\w+]+)$/;
         const nextRegex = /@(alife|alifd)\/next\/(es|lib)\/([-\w+]+)$/;
         const baseRegex = /@icedesign\/base\/lib\/([-\w+]+)$/;
-        externals.push(function(_context, request, callback) {
+        externals.push(function(_context: string, request: string, callback: Function) {
           const isNext = nextRegex.test(request);
           const isDesignBase = baseRegex.test(request);
           if (isNext || isDesignBase) {
@@ -393,3 +83,5 @@ module.exports = async ({ onGetWebpackConfig, log, context, getAllTask }, plugio
     });
   });
 };
+
+export default plugin;
