+++ tsconfig.json
@@ -39,5 +39,7 @@
     { "path": "packages/webpack-dev-mock" },
     { "path": "packages/plugin-keep-alive" },
     { "path": "packages/plugin-dev-inspector" },
+    { "path": "packages/plugin-fusion" },
+    { "path": "packages/plugin-fusion-legacy" },
   ]
 }
